# app/openapi.yaml
openapi: 3.1.0
info:
  title: CareerProxy API
  version: 1.0.0
  description: Job application consultancy platform API
servers:
  - url: https://careerproxy.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Resume:
      type: object
      properties:
        id: { type: string, format: uuid }
        candidate_id: { type: string, format: uuid }
        file_url: { type: string }
        file_name: { type: string }
        file_type: { type: string, enum: [pdf, docx] }
        is_active: { type: boolean }
        uploaded_at: { type: string, format: date-time }
...
    SubscriptionPlan:
      type: string
      enum: [starter, pro, premium]
...
security:
  - BearerAuth: []
paths:
...
  /resumes:
    get:
      summary: List candidate resumes
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Resume' }
    post:
      summary: Upload resume
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resume' }
        '413': { description: File too large }
        '422': { description: Invalid file type }
  /resumes/{id}:
    delete:
      summary: Delete resume
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
...
  /subscriptions/checkout:
    post:
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan: { $ref: '#/components/schemas/SubscriptionPlan' }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkout_url: { type: string }
...
  /webhooks/stripe:
    post:
      summary: Handle Stripe webhooks
      security: []
      responses:
        '200': { description: OK }
        '400': { description: Invalid signature }
...